day20 p2

+++ src/main/java/com/raja/tmp/day20/Number.java
@@ -1,8 +1,8 @@
 package com.raja.tmp.day20;
 
-public record Number(int id, int value) {
+public record Number(int id, long value) {
 
-    public static Number aNumber(int id, int value) {
+    public static Number aNumber(int id, long value) {
         return new Number(id, value);
     }
 }
